requirements.txt
absl-py==1.4.0
anyio==3.6.2
apache-beam==2.44.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==3.0.0
astunparse==1.6.3
attrs==21.4.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==4.2.4
certifi==2022.9.24
cffi==1.17.1
charset-normalizer==3.4.2
click==7.1.2
cloudpickle==2.2.0
colorama==0.4.6
comm==0.2.2
cramjam==2.10.0
crcmod==1.7
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
docker==4.4.4
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
facets-overview==1.1.1
fastavro==1.7.0
fasteners==0.18
fastjsonschema==2.16.2
Flask==2.0.3
flatbuffers==23.1.4
fqdn==1.5.1
gast==0.4.0
google-api-core==1.32.0
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.17.1
google-cloud-bigquery==2.34.4
google-cloud-bigquery-storage==2.13.2
google-cloud-bigtable==1.7.3
google-cloud-core==2.3.2
google-cloud-dataproc==3.1.1
google-cloud-datastore==1.15.5
google-cloud-dlp==3.9.2
google-cloud-language==1.3.2
google-cloud-pubsub==2.13.11
google-cloud-pubsublite==1.6.0
google-cloud-recommendations-ai==0.7.1
google-cloud-resource-manager==1.6.3
google-cloud-spanner==3.26.0
google-cloud-storage==2.7.0
google-cloud-videointelligence==1.16.3
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.48.2
h5py==3.7.0
hdfs==2.7.0
httplib2==0.20.4
idna==3.10
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.20.1
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.12.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter-latex-envs==1.4.6
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
jupyterlab_widgets==3.0.15
kagglehub==0.3.12
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kt-legacy==1.0.4
kubernetes==12.0.1
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
ml-metadata==1.11.0
ml-pipelines-sdk==1.11.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
opt-einsum==3.3.0
orjson==3.8.5
overrides==6.5.0
packaging==20.9
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
portpicker==1.5.2
prometheus-client==0.15.0
prompt_toolkit==3.0.51
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydot==1.4.2
pyfarmhash==0.3.2
Pygments==2.14.0
pymongo==3.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python-snappy==0.7.3
pytz==2022.7.1
pywin32==227
pywinpty==2.0.15
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-data-validation==1.11.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow-metadata==1.11.0
tensorflow-model-analysis==0.42.0
tensorflow-serving-api==2.10.1
tensorflow-transform==1.11.0
termcolor==2.2.0
terminado==0.17.1
tfx==1.11.0
tfx-bsl==1.11.0
threadpoolctl==3.1.0
timeloop==1.0.2
tinycss2==1.2.1
tomli==2.2.1
tornado==6.2
tqdm==4.67.1
traitlets==5.8.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.6.0
uri-template==1.2.0
uritemplate==3.0.1
urllib3==2.4.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.14
wrapt==1.14.1
zipp==3.11.0
zstandard==0.19.0
